--- app/src/main/java/com/hasib/carebear/doctor/listener/ChamberAddingDialogTimeSetListener.java
@@ -1,5 +1,0 @@
-package com.hasib.carebear.doctor.listener;
-
-public interface ChamberAddingDialogTimeSetListener {
-    void setTime(String time);
-}
