--- app/src/main/java/com/hasib/carebear/doctor/listener/ProfileDetailsListener.java
@@ -1,7 +1,0 @@
-package com.hasib.carebear.doctor.listener;
-
-import com.hasib.carebear.doctor.container.UserDetails;
-
-public interface ProfileDetailsListener {
-    void shareUserProfileData(UserDetails userDetails);
-}
