--- app/src/main/java/com/hasib/carebear/doctor/listener/TimePickerListener.java
@@ -1,7 +1,0 @@
-package com.hasib.carebear.doctor.listener;
-
-import android.widget.TimePicker;
-
-public interface TimePickerListener {
-    void onTimeSet(TimePicker timePicker, int hour, int minute);
-}
