--- app/src/main/java/com/hasib/carebear/support/ImageSupport.java
@@ -1,57 +1,0 @@
-package com.hasib.carebear.support;
-
-import android.graphics.Bitmap;
-import android.graphics.BitmapFactory;
-import android.os.AsyncTask;
-import android.util.Base64;
-
-import java.io.ByteArrayOutputStream;
-import java.io.IOException;
-import java.io.InputStream;
-import java.net.HttpURLConnection;
-import java.net.URL;
-
-public class ImageSupport {
-    public static Bitmap stringToBitMap(String encodedString){
-        try {
-            byte [] encodeByte = Base64.decode(encodedString, Base64.DEFAULT);
-            Bitmap bitmap = BitmapFactory.decodeByteArray(encodeByte, 0, encodeByte.length);
-            return bitmap;
-        } catch(Exception e) {
-            e.getMessage();
-            return null;
-        }
-    }
-
-    public static String bitMapToString(Bitmap bitmap){
-        ByteArrayOutputStream byteArrayOutputStream =new  ByteArrayOutputStream();
-        bitmap.compress(Bitmap.CompressFormat.PNG,100, byteArrayOutputStream);
-        byte [] b=byteArrayOutputStream.toByteArray();
-        String temp= Base64.encodeToString(b, Base64.URL_SAFE);
-        return temp;
-    }
-
-    private static class DownloadImageTask extends AsyncTask<String, Void, Bitmap> {
-        @Override
-        protected Bitmap doInBackground(String... strings) {
-            Bitmap bitmap=null;
-            URL url;
-            HttpURLConnection httpURLConnection=null;
-
-            try {
-                url = new URL(strings[0]);
-                httpURLConnection = (HttpURLConnection) url.openConnection();
-                httpURLConnection.connect();
-
-                InputStream inputStream = httpURLConnection.getInputStream();
-                bitmap = BitmapFactory.decodeStream(inputStream);
-
-                return bitmap;
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-
-            return null;
-        }
-    }
-}
