+++ app/src/main/java/com/tinvodai/carebear/doctor/listener/ChamberAddingDialogTimeSetListener.java
@@ -1,0 +1,5 @@
+package com.tinvodai.carebear.doctor.listener;
+
+public interface ChamberAddingDialogTimeSetListener {
+    void setTime(String time);
+}
