+++ app/src/main/java/com/tinvodai/carebear/doctor/listener/ProfileDetailsListener.java
@@ -1,0 +1,7 @@
+package com.tinvodai.carebear.doctor.listener;
+
+import com.tinvodai.carebear.doctor.container.UserDetails;
+
+public interface ProfileDetailsListener {
+    void shareUserProfileData(UserDetails userDetails);
+}
