+++ app/src/main/java/com/tinvodai/carebear/doctor/listener/TimePickerListener.java
@@ -1,0 +1,7 @@
+package com.tinvodai.carebear.doctor.listener;
+
+import android.widget.TimePicker;
+
+public interface TimePickerListener {
+    void onTimeSet(TimePicker timePicker, int hour, int minute);
+}
